serialization and optimize

+++ breakout/entitytype.nim
@@ -10,8 +10,8 @@ const
   maxEntities* = indexMask
 
 template idx*(e: Entity): int = e.int and indexMask
-template version*(e: Entity): int = e.int shr indexBits and versionMask
-template toEntity*(idx, v: int): Entity = Entity(v shl indexBits or idx)
+template version*(e: Entity): uint16 = e.uint16 shr indexBits and versionMask
+template toEntity*(idx, v: uint16): Entity = Entity(v shl indexBits or idx)
 
 proc `==`*(a, b: Entity): bool {.borrow.}
 proc `$`*(e: Entity): string =
